feat!: move issuer into oidcConfig
As mentioned:
> In general, when moving to support the OIDC discovery endpoint,
we would do it by default without being able to opt-out.
So all the configuration around being able to enable it,
configuring the issuer, endpoints, etc should just be removed
as we'll pull this from the doc.

Moving issuer from the options to options.oidcConfig
will break current configs.

In that case, this is a
BREAKING CHANGE